Prevent Firefox from remembering input type
By default, Firefox remembers form values across refreshes [0]. This
means the date input type selection could get out of sync.

I could run the update function when the page loads to sync with
whatever's in the select, but I think it makes more sense to not reload
that value so the user always starts with the first option.

[0]: https://stackoverflow.com/questions/5985839/bug-with-firefox-disabled-attribute-of-input-not-resetting-when-refreshing